requirements.txt
accelerate
aiohttp
aiosignal
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
async-timeout
attrs
babel
beautifulsoup4
black
bleach
certifi
cffi
cfgv
charset-normalizer
click
comm
cryptography
debugpy
decorator
defusedxml
distlib
et-xmlfile
exceptiongroup
executing
fastjsonschema
filelock
fqdn
frozenlist
fsspec
h11
httpcore
httpx
huggingface-hub
identify
idna
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
MarkupSafe
matplotlib-inline
mistune
mpmath
multidict
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nodeenv
notebook
notebook_shim
numpy
openpyxl
overrides
packaging
pandas
pandocfilters
parso
pathspec
pexpect
platformdirs
pre_commit
prometheus_client
prompt_toolkit
propcache
psutil
ptyprocess
pure_eval
pycparser
Pygments
PyJWT
python-dateutil
python-dotenv
python-json-logger
pytz
PyYAML
pyzmq
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rpds-py
safetensors
scikit-learn
scipy
Send2Trash
six
sniffio
soupsieve
stack-data
sympy
terminado
threadpoolctl
tinycss2
tokenize-rt
tokenizers
tomli
torch
tornado
tqdm
traitlets
transformers
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
virtualenv
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
yandexgpt-python
yarl
